Split monolithic Acceptance helper class into multiple classes by functionality
