+++ tests/_support/Helper/Acceptance/ConvertKitAPI.php
@@ -1,0 +1,91 @@
+<?php
+namespace Helper\Acceptance;
+
+// Define any custom actions related to the ConvertKit API that
+// would be used across multiple tests.
+// These are then available in $I->{yourFunctionName}
+
+class ConvertKitAPI extends \Codeception\Module
+{
+	/**
+	 * Check the given email address exists as a subscriber.
+	 * 
+	 * @param 	AcceptanceTester $I 			AcceptanceTester
+	 * @param 	string 			$emailAddress 	Email Address
+	 */ 	
+	public function apiCheckSubscriberExists($I, $emailAddress)
+	{
+		// Run request.
+		$results = $this->apiRequest('subscribers', 'GET', [
+			'email_address' => $emailAddress,
+		]);
+
+		// Check at least one subscriber was returned and it matches the email address.
+		$I->assertGreaterThan(0, $results['total_subscribers']);
+		$I->assertEquals($emailAddress, $results['subscribers'][0]['email_address']);
+	}
+
+	/**
+	 * Check the given email address does not exists as a subscriber.
+	 * 
+	 * @param 	AcceptanceTester $I 			AcceptanceTester
+	 * @param 	string 			$emailAddress 	Email Address
+	 */ 	
+	public function apiCheckSubscriberDoesNotExist($I, $emailAddress)
+	{
+		// Run request.
+		$results = $this->apiRequest('subscribers', 'GET', [
+			'email_address' => $emailAddress,
+		]);
+
+		// Check no subscribers are returned by this request.
+		$I->assertEquals(0, $results['total_subscribers']);
+	}
+
+	/**
+	 * Unsubscribes the given email address. Useful for clearing the API
+	 * between tests.
+	 * 
+	 * @param 	string 			$emailAddress 	Email Address
+	 */ 	
+	public function apiUnsubscribe($emailAddress)
+	{
+		// Run request.
+		$this->apiRequest('unsubscribe', 'PUT', [
+			'email' => $emailAddress,
+		]);
+	}
+
+	/**
+	 * Sends a request to the ConvertKit API, typically used to read an endpoint to confirm
+	 * that data in an Acceptance Test was added/edited/deleted successfully.
+	 * 
+	 * @param 	string 	$endpoint 	Endpoint
+	 * @param 	string 	$method 	Method (GET|POST|PUT)
+	 * @param 	array 	$params 	Endpoint Parameters
+	 */
+	public function apiRequest($endpoint, $method = 'GET', $params = array())
+	{
+		// Build query parameters.
+		$params = array_merge($params, [
+			'api_key' => $_ENV['CONVERTKIT_API_KEY'],
+			'api_secret' => $_ENV['CONVERTKIT_API_SECRET'],
+		]);
+
+		// Send request.
+		try {
+			$client = new \GuzzleHttp\Client();
+			$result = $client->request($method, 'https://api.convertkit.com/v3/' . $endpoint . '?' . http_build_query($params), [
+				'headers' => [
+					'Accept-Encoding' => 'gzip',
+					'timeout'         => 5,
+				],
+			]);
+
+			// Return JSON decoded response.
+			return json_decode($result->getBody()->getContents(), true);
+		} catch(\GuzzleHttp\Exception\ClientException $e) {
+			return [];
+		}
+	}
+}
