+++ tests/acceptance/WishListMemberCest.php
@@ -20,7 +20,7 @@ class WishListMemberCest
 		$I->setupConvertKitPlugin($I);
 		$I->enableDebugLog($I);
 		$I->wait(2);
-		$I->setupWishListMemberPlugin($I);
+		$I->_setupWishListMemberPlugin($I);
 	}
 
 	/**
@@ -111,6 +111,44 @@ class WishListMemberCest
 			'last_name' => 'User',
 			'display_name' => 'Test User',
 		]);
+	}
+
+	/**
+	 * Helper method to setup the WishList Member Plugin.
+	 * 
+	 * @since 	1.9.6
+	 */
+	private function _setupWishListMemberPlugin($I)
+	{
+		// Load WishList Member Settings screen, which will load the first time configuration wizard.
+		$I->amOnAdminPage('admin.php?page=WishListMember');
+
+		// Skip Licensing
+		$I->click('a.skip-license');
+		$I->performOn( 'a[next-screen="start"]', function($I) {
+			$I->click('a[next-screen="start"]');
+		});
+		
+		// Step 1
+		$I->fillField('input[name="name"]', 'Bronze');
+		$I->click('.step-1 a[next-screen="step-2"]');
+
+		// Step 2
+		$I->click('.step-2 a[next-screen="step-3"]');
+
+		// Step 3
+		$I->click('.step-3 a[next-screen="step-4"]');
+
+		// Step 4
+		$I->click('.step-4 a[next-screen="step-5"]');
+
+		// Save
+		$I->click('.step-5 a.save-btn');
+
+		$I->performOn('.-congrats-gs', function($I) {
+			$I->click('a.next-btn');
+			$I->seeInSource('Bronze');
+		});
 	}
 
 	/**
